+++ R/theme61-options.R
@@ -7,7 +7,7 @@
 #'
 #' @details The following options are available to set:
 #'  \itemize{
-#'    \item \code{theme61.open_e61_graph}: If TRUE, graphs will open in the browser instead of the Viewer pane. This is FALSE by default.
+#'    \item \code{theme61.open_in_browser}: If TRUE, graphs will open in the browser instead of the Viewer pane. This is FALSE by default.
 #'    \item \code{theme61.default_save_format}: The default file save format if format is not specified in [save_e61] and the file extension is not provided in \code{filename}. This is "svg" by default.
 #'    \item \code{theme61.preview_on_print}: If TRUE (default), graphs will be automatically previewed in the Viewer pane when printed to the console.
 #'    \item \code{theme61.base_size}: The base font size for graphs. This is 10 by default.
@@ -65,7 +65,7 @@ set_t61_options <- function(opt = NULL) {
 #' @rdname open_graph_browser
 #' @export
 set_open_graph_browser <- function() {
-  options(theme61.open_e61_graph = TRUE)
+  options(theme61.open_in_browser = TRUE)
 
   invisible(TRUE)
 }
@@ -73,7 +73,7 @@ set_open_graph_browser <- function() {
 #' @rdname open_graph_browser
 #' @export
 unset_open_graph_browser <- function() {
-  options(theme61.open_e61_graph = FALSE)
+  options(theme61.open_in_browser = FALSE)
 
   invisible(FALSE)
 }