+++ R/zzz.R
@@ -43,7 +43,7 @@ t61_env <- NULL
     theme61.base_size = 10,
     theme61.default_save_format = "svg",
     theme61.iterate_mode = FALSE,
-    theme61.open_e61_graph = FALSE,
+    theme61.open_in_browser = FALSE,
     theme61.preview_on_print = TRUE
   )